+++ lib/app.js
@@ -6,13 +6,11 @@ app.use(morgan('dev'));
 // app.use(express.static('public'));
 app.use(express.json());
 
-const ensureAuth = require('./util/ensure-auth')();
+// const ensureAuth = require('./util/ensure-auth')();
 
 const reviewers = require('./routes/reviewers');
-const auth = require('./routes/auth');
 
-app.use('/api/auth', auth);
-app.use('/api/reviewers', ensureAuth, reviewers);
+app.use('/api/reviewers', reviewers);
 
 const studios = require('./routes/studios');
 app.use('/api/studios', studios);
